Refactored Java YAML parser out of test class
